Collect all unique quest filters and implement it #68 - filter
