+++ src/BlScraper.DependencyInjection/ConfigureBuilder/ScrapBuilderConfig.cs
@@ -1,3 +1,4 @@
+using BlScraper.DependencyInjection.Builder.Internal;
 using BlScraper.DependencyInjection.ConfigureModel.Filter;
 using System.Collections;
 
@@ -13,6 +14,9 @@ public class ScrapBuilderConfig
     /// </summary>
     private readonly PoolFilter _poolFilters = new();
 
+    /// <inheritdoc cref="_poolFilters" path="*"/>
+    internal PoolFilter Filters => _poolFilters;
+
     /// <summary>
     /// Assemblies to map models
     /// </summary>
@@ -22,42 +26,6 @@ public class ScrapBuilderConfig
     /// Assemblies to map models
     /// </summary>
     internal IEnumerable<System.Reflection.Assembly> Assemblies => _assemblies;
-
-    /// <summary>
-    /// <see cref="IAllWorksEndConfigureFilter"/> pool
-    /// </summary>
-    internal IEnumerable<Type> PoolAllWorksEndConfigureFilter =>
-        _poolFilters.Where(tuple => typeof(IAllWorksEndConfigureFilter).IsAssignableFrom(tuple.FilterInterface)).Select(tuple => tuple.Filter);
-
-    /// <summary>
-    /// <see cref="IDataCollectedConfigureFilter"/> pool
-    /// </summary>
-    internal IEnumerable<Type> PoolDataCollectedConfigureFilter =>
-        _poolFilters.Where(tuple => typeof(IDataCollectedConfigureFilter).IsAssignableFrom(tuple.FilterInterface)).Select(tuple => tuple.Filter);
-
-    /// <summary>
-    /// <see cref="IDataFinishedConfigureFilter"/> pool
-    /// </summary>
-    internal IEnumerable<Type> PoolDataFinishedConfigureFilter =>
-        _poolFilters.Where(tuple => typeof(IDataFinishedConfigureFilter).IsAssignableFrom(tuple.FilterInterface)).Select(tuple => tuple.Filter);
-    
-    /// <summary>
-    /// <see cref="IGetArgsConfigureFilter"/> pool
-    /// </summary>
-    internal IEnumerable<Type> PoolGetArgsConfigureFilter =>
-        _poolFilters.Where(tuple => typeof(IGetArgsConfigureFilter).IsAssignableFrom(tuple.FilterInterface)).Select(tuple => tuple.Filter);
-
-    /// <summary>
-    /// <see cref="IQuestCreatedConfigureFilter"/> pool
-    /// </summary>
-    internal IEnumerable<Type> PoolQuestCreatedConfigureFilter =>
-        _poolFilters.Where(tuple => typeof(IQuestCreatedConfigureFilter).IsAssignableFrom(tuple.FilterInterface)).Select(tuple => tuple.Filter);
-
-    /// <summary>
-    /// <see cref="IQuestExceptionConfigureFilter"/> pool
-    /// </summary>
-    internal IEnumerable<Type> PoolQuestExceptionConfigureFilter =>
-        _poolFilters.Where(tuple => typeof(IQuestExceptionConfigureFilter).IsAssignableFrom(tuple.FilterInterface)).Select(tuple => tuple.Filter);
 
     /// <summary>
     /// Lock object
@@ -102,10 +70,11 @@ public class ScrapBuilderConfig
     /// Add filter of type <see cref="IAllWorksEndConfigureFilter"/>
     /// </summary>
     /// <param name="filter">filter to add</param>
-    /// <inheritdoc cref="CheckAndAddToPoll(Type, Type)" path="/exception"/>
+    /// <inheritdoc cref="PoolFilter.Add(Type, Type)" path="/exception"/>
     public ScrapBuilderConfig AddAllWorksEndConfigureFilter(Type filter)
     {
-        return CheckAndAddToPoll(typeof(IAllWorksEndConfigureFilter), filter);
+        _poolFilters.Add(typeof(IAllWorksEndConfigureFilter), filter);
+        return this;
     }
 
     /// <summary>
@@ -123,10 +92,11 @@ public class ScrapBuilderConfig
     /// Add filter of type <see cref="IDataCollectedConfigureFilter"/>
     /// </summary>
     /// <param name="filter">filter to add</param>
-    /// <inheritdoc cref="CheckAndAddToPoll(Type, Type)" path="/exception"/>
+    /// <inheritdoc cref="PoolFilter.Add(Type, Type)" path="/exception"/>
     public ScrapBuilderConfig AddDataCollectedConfigureFilter(Type filter)
     {
-        return CheckAndAddToPoll(typeof(IDataCollectedConfigureFilter), filter);
+        _poolFilters.Add(typeof(IDataCollectedConfigureFilter), filter);
+        return this;
     }
 
     /// <summary>
@@ -137,17 +107,19 @@ public class ScrapBuilderConfig
     public ScrapBuilderConfig AddDataFinishedConfigureFilter<TFilter>()
         where TFilter : IDataFinishedConfigureFilter
     {
-        return AddDataFinishedConfigureFilter(typeof(TFilter));
+        AddDataFinishedConfigureFilter(typeof(TFilter));
+        return this;
     }
 
     /// <summary>
     /// Add filter of type <see cref="IDataFinishedConfigureFilter"/>
     /// </summary>
     /// <param name="filter">filter to add</param>
-    /// <inheritdoc cref="CheckAndAddToPoll(Type, Type)" path="/exception"/>
+    /// <inheritdoc cref="PoolFilter.Add(Type, Type)" path="/exception"/>
     public ScrapBuilderConfig AddDataFinishedConfigureFilter(Type filter)
     {
-        return CheckAndAddToPoll(typeof(IDataFinishedConfigureFilter), filter);
+        _poolFilters.Add(typeof(IDataFinishedConfigureFilter), filter);
+        return this;
     }
     
     /// <summary>
@@ -165,10 +137,11 @@ public class ScrapBuilderConfig
     /// Add filter of type <see cref="IGetArgsConfigureFilter"/>
     /// </summary>
     /// <param name="filter">filter to add</param>
-    /// <inheritdoc cref="CheckAndAddToPoll(Type, Type)" path="/exception"/>
+    /// <inheritdoc cref="PoolFilter.Add(Type, Type)" path="/exception"/>
     public ScrapBuilderConfig AddGetArgsConfigureFilter(Type filter)
     {
-        return CheckAndAddToPoll(typeof(IGetArgsConfigureFilter), filter);
+        _poolFilters.Add(typeof(IGetArgsConfigureFilter), filter);
+        return this;
     }
 
     /// <summary>
@@ -179,17 +152,19 @@ public class ScrapBuilderConfig
     public ScrapBuilderConfig AddQuestCreatedConfigureFilter<TFilter>()
         where TFilter : IQuestCreatedConfigureFilter
     {
-        return AddQuestCreatedConfigureFilter(typeof(TFilter));
+        AddQuestCreatedConfigureFilter(typeof(TFilter));
+        return this;
     }
 
     /// <summary>
     /// Add filter of type <see cref="IQuestCreatedConfigureFilter"/>
     /// </summary>
     /// <param name="filter">filter to add</param>
-    /// <inheritdoc cref="CheckAndAddToPoll(Type, Type)" path="/exception"/>
+    /// <inheritdoc cref="PoolFilter.Add(Type, Type)" path="/exception"/>
     public ScrapBuilderConfig AddQuestCreatedConfigureFilter(Type filter)
     {
-        return CheckAndAddToPoll(typeof(IQuestCreatedConfigureFilter), filter);
+        _poolFilters.Add(typeof(IQuestCreatedConfigureFilter), filter);
+        return this;
     }
 
     /// <summary>
@@ -207,104 +182,10 @@ public class ScrapBuilderConfig
     /// Add filter of type <see cref="IQuestExceptionConfigureFilter"/>
     /// </summary>
     /// <param name="filter">filter to add</param>
-    /// <inheritdoc cref="CheckAndAddToPoll(Type, Type)" path="/exception"/>
+    /// <inheritdoc cref="PoolFilter.Add(Type, Type)" path="/exception"/>
     public ScrapBuilderConfig AddQuestExceptionConfigureFilter(Type filter)
     {
-        return CheckAndAddToPoll(typeof(IQuestExceptionConfigureFilter), filter);
-    }
-
-    /// <summary>
-    /// Private check and add
-    /// </summary>
-    /// <param name="filterInterface">filter interface</param>
-    /// <param name="filter">filter implementation</param>
-    /// <inheritdoc cref="CheckAndThrowFilter(Type, Type)" path="/exception"/>
-    private ScrapBuilderConfig CheckAndAddToPoll(Type filterInterface, Type filter)
-    {
-        CheckAndThrowFilter(filter, filterInterface);
-
-        _poolFilters.Add(filterInterface, filter);
-
+        _poolFilters.Add(typeof(IQuestExceptionConfigureFilter), filter);
         return this;
-    }
-
-    /// <summary>
-    /// Check options of filter and if is assgnable to <paramref name="assignable"/>
-    /// </summary>
-    /// <param name="filter">filter</param>
-    /// <param name="assignable">filter assignable</param>
-    /// <exception cref="ArgumentNullException"></exception>
-    /// <exception cref="ArgumentException"></exception>
-    private static void CheckAndThrowFilter(Type filter, Type assignable)
-    {
-        if (filter is null)
-            throw new ArgumentNullException($"filter");
-
-        if (assignable is null)
-            throw new ArgumentNullException($"assignable");
-
-        if (!assignable.IsAssignableFrom(filter))
-            throw new ArgumentException($"Filter: '{filter.FullName}' isn't assignable to '{assignable.FullName}'.", $"{filter.FullName}");
-
-        if (!filter.IsClass)
-            throw new ArgumentException($"Filter: '{filter.FullName}' must be a class.", $"{filter.FullName}");
-
-        if (filter.IsAbstract)
-            throw new ArgumentException($"Filter: '{filter.FullName}' must not be abstract.", $"{filter.FullName}");
-
-        if (!filter.IsPublic)
-            throw new ArgumentException($"Filter: '{filter.FullName}' must be public.", $"{filter.FullName}");
-
-        if (filter.ContainsGenericParameters)
-            throw new ArgumentException($"Filter: '{filter.FullName}' must not have generic parameters.", $"{filter.FullName}");
-    }
-    
-    /// <summary>
-    /// Sync Pool filter
-    /// </summary>
-    private class PoolFilter : IEnumerable<(Type FilterInterface, Type Filter)>
-    {
-        /// <summary>
-        /// Type filters pool
-        /// </summary>
-        private System.Collections.Concurrent.ConcurrentBag<(Type filterInterface, Type Filter)> _poolFilters = new();
-
-        /// <summary>
-        /// Try add
-        /// </summary>
-        /// <param name="filterInterface">Filter interface</param>
-        /// <param name="filter">Filter</param>
-        /// <returns>true : item added, otherwise already contains</returns>
-        public bool TryAdd(Type filterInterface, Type filter)
-        {
-            if (_poolFilters.Contains((filterInterface, filter)))
-                return false;
-
-            _poolFilters.Add((filterInterface, filter));
-
-            return true;
-        }
-
-        /// <summary>
-        /// Add and check
-        /// </summary>
-        /// <param name="filterInterface">Filter interface</param>
-        /// <param name="filter">Filter</param>
-        /// <exception cref="ArgumentException"/>
-        public void Add(Type filterInterface, Type filter)
-        {
-            if (!TryAdd(filterInterface, filter))
-                throw new ArgumentException($"List already contains '{filter.FullName}' to '{filterInterface.FullName}'.", filter.FullName);
-        }
-
-        public IEnumerator<(Type FilterInterface, Type Filter)> GetEnumerator()
-        {
-            return _poolFilters.GetEnumerator();
-        }
-
-        IEnumerator IEnumerable.GetEnumerator()
-        {
-            return GetEnumerator();
-        }
     }
 }
